+++ app/components/Navbar.tsx
@@ -6,7 +6,7 @@ import Link from "next/link";
 import fluenturesLogo from "@/fluenturesLogo.png";
 import Image from "next/image";
 import { Button } from "@/app/components/Button";
-import { supabase } from '@/app/lib/clients/supabaseClient'
+import { supabase } from '@/app/lib/actions/supabaseClient'
 import { useRouter } from 'next/navigation';
 
 const Navbar = () => {
