+++ app/oasis/quiz/WrittenQuiz.tsx
@@ -2,7 +2,7 @@
 
 import { Button } from "@/app/components/Button";
 import { useEffect, useState } from "react";
-import { requestQuizSentence } from "@/app/lib/clients/geminiQuizClient";
+import { requestQuizSentence } from "@/app/lib/actions/geminiQuizClient";
 
 // Simulated word list (target language: Japanese)
 const wordList = [
