+++ pages/api/signup.ts
@@ -1,5 +1,5 @@
 import type { NextApiRequest, NextApiResponse } from 'next'
-import { supabase } from '@/app/lib/clients/supabaseAdminClient'
+import { supabase } from '@/app/lib/actions/supabaseAdminClient'
 
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
